Update `requirements.txt` and tests to get them passing without being skipped again (#42)
* Bump requirements.txt by running pip-compile --generate-hashes
  * we had to manually add hashes back for `async-timeout`

* Set asyncio_mode=True

* For aiohttp_session, remove loop fixture and mark as async

`loop` is deprecated, but as I was switching to the new fixture, I found that it wasn't needed at all

* Use mock_cert helper in test_verify_x5u_screwy_dates test

* Create mock cert with `wraps`, rather than `spec`

This mock object used to be created with `mock.create_autospec(spec=real_cert)`. The certs we use for testing don't have SCT(Signed Certificate Timestamp) data in them. So when Python tried to create the spec, it tried to access a property that throws a ValueError if a certificate doesn't have that data

https://github.com/pyca/cryptography/blob/0b2a62f77db6ad1c2d9297ec3c4569aea4712a43/tests/x509/test_x509.py#L1154-L1178

By switching from `spec` to `wraps`, we avoid this error and maintain the same safety as using spec. Quoting the `unittest.mock` docs:

> If wraps is not None then calling the Mock will pass the call through to the wrapped object (returning the real result). Attribute access on the mock will return a Mock object that wraps the corresponding attribute of the wrapped object (so attempting to access an attribute that doesn’t exist will raise an AttributeError).

* Explicitly set python-version for setup-python in lint job

also, enable caching